major reunderstanding of StartCMD
Previously confuse dby the doubling of the "A" start cmd for "write
text file" in the docs to "AA" in working example code.  Second "A" is
a "File Label" here's how th e documentation defines that:

  A File Label is a single ASCII character.  Messages are stored in or
  retrieved from the memory file that is defined by this label in the
  MEMORY CONFIGURATION.  Legal File Labels can be anywhere in the
  range "sp" (20H) thru "res" (7FH) inclusive.  The only special case
  occurs when File Label "0" (30H) is used for a TEXT file.  This is
  an illegal label for a TEXT file in the MEMORY CONFIGURATION.  It is
  already configured as a set portion of memory outside of the MEMORY
  POOL, as a PRIORITY TEXT file.  See the section on PRIORITY TEXT
  file for further information

So there are is a moderate range of storable addresses.  For now just
ackowlege this is variable at the instance level and leave it as acsii
"A" but untouchable for now.